Made a start on displacement maps implementation. Somehow need to hook into TextureNode to find connections. Probably need another "dummy" texture node displacementShader to handle the attributes and I/O. Use bumpmapTexture as a model.
Another interesting thing worth noting is that we have "useMaxSubdivisions" and "maxSubd" in the mesh node attributes which can be used to control trianglemesh/loopsubdiv output for polygon meshes and the nlevels parameter. It's disabled by default. -- the optimised mesh parser needs to work 100% before this can be implemented though.